stage1.sh: They changed the pacman download technology (possibly when parallel download was implemented). It seems that the 1st server that works will be used for the rest of the downloads.
This is a problem: In the past I prepended the mirror list with web
server on my local /var/cache/pacman/pkg, which obviously fails on the
pacman database refresh. Hence, it will not be used for the download of
normal packages, either, now. This worked before as the download always
retried all the hosts in the mirrorlist in each pacman invocation.

However, this is a lot better solution, and I have to adapt. So I had to
change to using pacserve by https://xyne.dev/projects/pacserve/

Also a small bugfix on initializing the country.